Added trigger for actor runs
