+++ package.json
@@ -7,6 +7,7 @@
   "main": "index.js",
   "scripts": {
     "test": "mocha --recursive",
+    "test:grep": "mocha --recursive --grep \"${GREP}\"",
     "lint": "./node_modules/.bin/eslint ./src index.js --ext .js",
     "lint:fix": "./node_modules/.bin/eslint ./src --ext .js --fix"
   },
