+++ src/apify_helpers.js
@@ -1,4 +1,5 @@
 const Promise = require('bluebird');
+const { WEBHOOK_EVENT_TYPES } = require('apify-shared/consts');
 const { APIFY_API_ENDPOINTS, DEFAULT_KEY_VALUE_STORE_KEYS } = require('./consts');
 const { wrapRequestWithRetries } = require('./request_helpers');
 
@@ -74,6 +75,42 @@ const enrichTaskRun = async (z, run, storeKeysToInclude = []) => {
     return run;
 };
 
+const subscribeWebkook = async (z, bundle, condition) => {
+    const webhookOpts = {
+        eventTypes: Object.values(WEBHOOK_EVENT_TYPES),
+        condition,
+        requestUrl: bundle.targetUrl,
+    };
+    const response = await wrapRequestWithRetries(z.request, {
+        url: APIFY_API_ENDPOINTS.webhooks,
+        method: 'POST',
+        json: webhookOpts,
+    });
+
+    return response.json;
+};
+
+const unsubscribeWebhook = async (z, bundle) => {
+    // bundle.subscribeData contains the parsed response JSON from the subscribe
+    const webhookId = bundle.subscribeData.id;
+
+    await wrapRequestWithRetries(z.request, {
+        url: `${APIFY_API_ENDPOINTS.webhooks}/${webhookId}`,
+        method: 'DELETE',
+    });
+
+    return {};
+};
+
+const getActorkRun = async (z, bundle) => {
+    const run = bundle.cleanedRequest.resource;
+    const enrichRun = await enrichTaskRun(z, run);
+    return [enrichRun];
+};
+
 module.exports = {
     enrichTaskRun,
+    subscribeWebkook,
+    unsubscribeWebhook,
+    getActorkRun,
 };
