+++ src/triggers/actors.js
@@ -1,32 +1,32 @@
 const { APIFY_API_ENDPOINTS, DEFAULT_PAGINATION_LIMIT } = require('../consts');
 const { wrapRequestWithRetries } = require('../request_helpers');
 
-// Fetches a list of tasks
-const getTaskList = async (z, bundle) => {
-    const taskListResponse = await wrapRequestWithRetries(z.request, {
-        url: `${APIFY_API_ENDPOINTS.tasks}`,
+// Fetches a list of actors
+const getActorList = async (z, bundle) => {
+    const actorListResponse = await wrapRequestWithRetries(z.request, {
+        url: `${APIFY_API_ENDPOINTS.actors}`,
         params: {
             limit: DEFAULT_PAGINATION_LIMIT,
             offset: bundle.meta.page ? bundle.meta.page * DEFAULT_PAGINATION_LIMIT : 0,
         },
     });
-    return taskListResponse.json.items.map((task) => ({
-        id: task.id,
-        name: task.name,
+    return actorListResponse.json.items.map((actor) => ({
+        id: actor.id,
+        name: actor.name,
     }));
 };
 
 module.exports = {
-    key: 'tasks',
-    noun: 'Tasks',
+    key: 'actors',
+    noun: 'Actors',
     display: {
-        label: 'List of Tasks',
+        label: 'List of actors',
         description: 'This is a hidden trigger, and is used in a Dynamic Dropdown within this app',
         hidden: true,
     },
     operation: {
         // since this is a "hidden" trigger, there aren't any inputFields needed
-        perform: getTaskList,
+        perform: getActorList,
         canPaginate: true,
     },
 };
